Moved WireMock mocks and logic to new WordPressMocks subproject
