+++ WordPress/build.gradle
@@ -197,7 +197,7 @@ dependencies {
         exclude group: 'com.android.support', module: 'support-v4'
         exclude module: 'recyclerview-v7'
     }
-    androidTestImplementation("com.github.tomakehurst:wiremock:2.21.0") {
+    androidTestImplementation('com.github.tomakehurst:wiremock:2.21.0') {
         exclude group: 'org.apache.httpcomponents', module: 'httpclient'
         exclude group: 'org.apache.commons', module: 'commons-lang3'
         exclude group: 'asm', module: 'asm'
