--- libs/mocks/WordPressMocks/src/main/assets/lib-assets/abc.txt
@@ -1,1 +1,0 @@
-Hello World!
