+++ libs/mocks/WordPressMocks/src/main/assets/mocks/mappings/self-hosted/xmlrpcphp-wp.getUsersBlogs.json
