+++ libs/mocks/WordPressMocks/src/main/assets/mocks/mappings/wpcom/rest_v11_me_settings.json
