+++ libs/mocks/WordPressMocks/src/main/assets/mocks/mappings/wpcom/rest_v11_sites_158396482_post-formats.json
