+++ libs/mocks/WordPressMocks/src/main/assets/mocks/mappings/wpcom/rest_v12_read_following_mine.json
