+++ libs/mocks/WordPressMocks/src/main/assets/mocks/mappings/wpcom/rest_v13_auth_send-login-email.json
