+++ libs/mocks/WordPressMocks/src/main/assets/mocks/mappings/wpcom/rest_v1_devices_id_delete.json
