--- libs/mocks/WordPressMocks/src/main/java/org/wordpress/android/mocks/TestClass.java
@@ -1,4 +1,0 @@
-package org.wordpress.android.mocks;
-
-public class TestClass {
-}
